reorg session tests
